exampleをディレクトリに分ける
