+++ example/step3/main.go
@@ -1,31 +1,32 @@
-// #4 天気確認機能の実装
 package main
 
 // 利用したい外部のコードを読み込む
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/joho/godotenv"
 	"log"
 	"math/rand"
+	"net"
 	"net/http"
 	"os"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/go-sql-driver/mysql"
+	"github.com/jmoiron/sqlx"
+	"github.com/joho/godotenv"
 	"github.com/line/line-bot-sdk-go/v7/linebot"
 )
 
+// main関数外で利用するためにここで宣言する
+// 詳しくは「スコープ」や「グローバル変数」で検索してください
+var (
+	db *sqlx.DB
+)
 
 // init関数はmain関数実行前の初期化のために呼び出されることがGo言語の仕様として決まっている
 func init() {
-	// ランダムな数値を生成する際のシード値の設定
-	rand.Seed(time.Now().UnixNano())
-}
-
-// main関数は最初に呼び出されることがGo言語の仕様として決まっている
-func main() {
 	// ここで.envファイル全体を読み込みます。
 	// この読み込み処理がないと、個々の環境変数が取得出来ません。
 	// 読み込めなかったら err にエラーが入ります。
@@ -36,6 +37,24 @@ func main() {
 		fmt.Printf("読み込み出来ませんでした: %v", err)
 	}
 
+	// ランダムな数値を生成する際のシード値の設定
+	rand.Seed(time.Now().UnixNano())
+	// データベースへ接続する
+	config := mysql.Config{
+		User:                 os.Getenv("DB_USERNAME"),
+		Passwd:               os.Getenv("DB_PASSWORD"),
+		Net:                  "tcp",
+		Addr:                 net.JoinHostPort(os.Getenv("DB_HOSTNAME"), os.Getenv("DB_PORT")),
+		DBName:               os.Getenv("DB_DATABASE"),
+		AllowNativePasswords: true,
+		ParseTime:            true,
+		Loc:                  time.Local,
+	}
+	db = sqlx.MustConnect("mysql", config.FormatDSN())
+}
+
+// main関数は最初に呼び出されることがGo言語の仕様として決まっている
+func main() {
 	// LINEのAPIを利用する設定
 	bot, err := linebot.New(
 		os.Getenv("CHANNEL_SECRET"),
@@ -62,7 +81,6 @@ func main() {
 			}
 			return
 		}
-
 		// LINEサーバから来たメッセージによって行う処理を変える
 		for _, event := range events {
 			switch event.Type {
@@ -74,7 +92,7 @@ func main() {
 				if _, err = bot.ReplyMessage(event.ReplyToken, linebot.NewTextMessage(replyMessage)).Do(); err != nil {
 					log.Print(err)
 				}
-			// それ以外
+			// それ以外のとき
 			default:
 				continue
 			}
@@ -95,12 +113,21 @@ const helpMessage = `使い方
 	スタンプの情報を答えるよ！
 位置情報:
 	その場所の天気・気温・湿度を答えるよ！
+TodoList:
+	"todo"に続けて実行したい操作を入力してね！
+		list
+		add "タスク名" "期限"
+		done "タスクID"
+	例:
+		todo list
+		todo add レポート 2/24
+		todo done 12
 それ以外:
 	それ以外にはまだ対応してないよ！ごめんね...`
 
 // 返信を生成する
 func getReplyMessage(event *linebot.Event) (replyMessage string) {
-	// 来たメッセージの種類によって行う処理を変える
+	// 来たメッセージの種類によって分岐する
 	switch message := event.Message.(type) {
 	// テキストメッセージが来たとき
 	case *linebot.TextMessage:
@@ -108,15 +135,21 @@ func getReplyMessage(event *linebot.Event) (replyMessage string) {
 		if strings.Contains(message.Text, "おみくじ") {
 			// おみくじ結果を取得する
 			return getFortune()
+			// あるいは「todo」という文字列で始まるとき
+		} else if strings.HasPrefix(message.Text, "todo") {
+			// Todo用のメッセージを生成する
+			return dealTodo(message)
 		}
+
 		// それ以外のときはオウム返しする
 		return message.Text
 
 	// スタンプが来たとき
 	case *linebot.StickerMessage:
-		return fmt.Sprintf("sticker id is %v, stickerResourceType is %v", message.StickerID, message.StickerResourceType)
+		replyMessage := fmt.Sprintf("sticker id is %v, stickerResourceType is %v", message.StickerID, message.StickerResourceType)
+		return replyMessage
 
-		// 位置情報が来たとき
+	// 位置情報が来たとき
 	case *linebot.LocationMessage:
 		// その場所の天気
 		replyMessage, err := getWeather(message)
@@ -145,7 +178,7 @@ func getFortune() string {
 		8: "中凶",
 		9: "大凶",
 	}
-	// rand.Intn(10)は1～10のランダムな整数を返す
+	// rand.Intn(10)は0～9のランダムな整数を返す
 	return oracles[rand.Intn(len(oracles))]
 }
 
@@ -172,7 +205,7 @@ func getWeather(location *linebot.LocationMessage) (string, error) {
 	// 緯度経度からOpenWeatherMapAPIのURLを作成
 	lat := strconv.FormatFloat(location.Latitude, 'f', 6, 64)
 	lon := strconv.FormatFloat(location.Longitude, 'f', 6, 64)
-	url := "https://api.openweathermap.org/data/2.5/weather?lat=" + lat + "&lon=" + lon + "&appid=" + os.Getenv("APP_ID")
+	url := fmt.Sprintf("http://api.openweathermap.org/data/2.5/weather?lat=%v&lon=%v&APPID=%v", lat, lon, os.Getenv("APP_ID"))
 
 	// OpenWeatherMapAPIへのリクエスト
 	res, err := http.Get(url)
@@ -195,4 +228,94 @@ func getWeather(location *linebot.LocationMessage) (string, error) {
 湿度 : ` + fmt.Sprintf("%.2f", weatherData.Info.Humidity) + "%"
 
 	return text, nil
+
+}
+
+// データベースでTodoを扱う形式 (構造体)
+type Task struct {
+	ID      uint   `db:"id"`
+	Todo    string `db:"todo"`
+	DueDate string `db:"due_date"`
+}
+
+// Todo用のメッセージを生成
+func dealTodo(message *linebot.TextMessage) string {
+	// 受け取ったメッセージを空白で区切る
+	token := strings.Split(message.Text, " ")
+
+	// 区切った文字列の個数が1つ以下のときはヘルプを返す
+	if len(token) <= 1 {
+		return helpMessage
+	}
+
+	// Todoリスト表示
+	if token[1] == "list" {
+		return getTodoList()
+		// TodoリストにTodoを追加
+	} else if token[1] == "add" {
+		return addTodo(token)
+		// Todoリストから指定したIDのTodoを削除
+	} else if token[1] == "done" {
+		return deleteTodo(token)
+	}
+	return helpMessage
+}
+
+// Todoリストの取得
+func getTodoList() string {
+	var tasks []Task
+	// MySQLデータベースへのクエリを発行して一覧を取得する
+	err := db.Select(&tasks, "SELECT * from tasks")
+	if err != nil {
+		fmt.Print(err)
+		return fmt.Sprintf("db error: %v", err)
+	}
+
+	// メッセージの生成
+	replyMessage := "ID/ToDo/期限"
+	for _, task := range tasks {
+		replyMessage += fmt.Sprintf("\n%d/%v/%v", task.ID, task.Todo, task.DueDate)
+	}
+	return replyMessage
+}
+
+// TodoリストへのTodoの追加
+func addTodo(token []string) string {
+	// MySQLデータベースへのクエリを発行してTodoを追加する
+	result, err := db.Exec("INSERT INTO tasks (todo, due_date) VALUES (?, ?)", token[2], token[3])
+	if err != nil {
+		log.Printf("db error: %v", err)
+		return "Botサーバーでエラーが発生しました"
+	}
+
+	// (最後の)追加されたTodoのIDを取得する
+	todoID, err := result.LastInsertId()
+	if err != nil {
+		log.Printf("db error: %v", err)
+		return "Botサーバーでエラーが発生しました"
+	}
+
+	// メッセージの生成
+	replyMessage := fmt.Sprintf("todo added\nID:%d\ntodo:%v\n期限:%v", todoID, token[2], token[3])
+	return replyMessage
+}
+
+// Todoの削除
+func deleteTodo(token []string) string {
+	// IDを文字列から数値に変換する
+	id, err := strconv.Atoi(token[2])
+	if err != nil {
+		return "Botサーバーでエラーが発生しました"
+	}
+
+	// MySQLデータベースへのクエリを発行してそのIDのTodoを削除する
+	_, err = db.Exec("DELETE FROM tasks WHERE id = ?", id)
+	if err != nil {
+		log.Printf("db error: %v", err)
+		return "Botサーバーでエラーが発生しました"
+	}
+
+	// メッセージの生成
+	replyMessage := fmt.Sprintf("todo deleted\nID:%d", id)
+	return replyMessage
 }
