+++ example/step4_advance/main.go
@@ -15,7 +15,6 @@ import (
 	"time"
 
 	"github.com/joho/godotenv"
-
 	"github.com/line/line-bot-sdk-go/v7/linebot"
 )
